Refactor client forms (#1352)
* Split client forms code base into different components

* Added disable client form btn functionality to utils

* Disable form save btn if it fails validation

* Moved to try catch & no reloading any more to update the client data

* Extracted project list on client details to table component

* Added TotalHoursChart component to show client total hours

* Extracted project table headers to constants

* Fixed client form system specs

* Removed unnecessry api error from client forms

* No more reloading on mobile client form update action
