+++ app/javascript/src/components/Clients/ClientForm/MobileClientForm.tsx
@@ -3,69 +3,36 @@ import React, { useEffect, useState } from "react";
 
 import { Country, State, City } from "country-state-city";
 import { Form, Formik, FormikProps } from "formik";
-import { XIcon, EditImageButtonSVG, deleteImageIcon } from "miruIcons";
+import { XIcon } from "miruIcons";
 import PhoneInput from "react-phone-number-input";
 import flags from "react-phone-number-input/flags";
-import { Avatar, Button, SidePanel, Toastr } from "StyledComponents";
+import { Button, SidePanel, Toastr } from "StyledComponents";
 
 import clientApi from "apis/clients";
 import CustomReactSelect from "common/CustomReactSelect";
 import { InputErrors, InputField } from "common/FormikFields";
 
 import { clientSchema, getInitialvalues } from "./formValidationSchema";
-import { formatFormData } from "./utils";
-
-import { i18n } from "../../../i18n";
-
-interface IClientForm {
-  clientLogoUrl: string;
-  handleDeleteLogo: any;
-  setClientLogoUrl: any;
-  setClientLogo: any;
-  formType?: string;
-  clientData?: any;
-  apiError?: string;
-  setClientData?: any;
-  setnewClient?: any;
-  clientLogo?: any;
-  setApiError?: any;
-  setShowEditDialog?: any;
-  submitting: any;
-  setSubmitting: any;
-  handleEdit?: any;
-  setShowDialog: any;
-}
-
-interface FormValues {
-  name: string;
-  email: string;
-  phone: string;
-  address1: string;
-  address2: string;
-  country: any;
-  state: any;
-  city: any;
-  zipcode: string;
-  logo: any;
-}
+import UploadLogo from "./UploadLogo";
+import { formatFormData, disableBtn } from "./utils";
 
 const MobileClientForm = ({
+  client,
   clientLogoUrl,
   handleDeleteLogo,
   setClientLogoUrl,
   setClientLogo,
   clientData,
   formType = "new",
-  apiError = "",
   setClientData,
   setnewClient,
   clientLogo,
-  setApiError,
   setShowEditDialog,
   submitting,
   setSubmitting,
   handleEdit,
   setShowDialog,
+  fetchDetails,
 }: IClientForm) => {
   const [fileUploadError, setFileUploadError] = useState<string>("");
   const [countries, setCountries] = useState([]);
@@ -110,38 +77,6 @@ const MobileClientForm = ({
     return cities;
   };
 
-  const onLogoChange = e => {
-    const file = e.target.files[0];
-    const isValid = isValidFileUploaded(file);
-
-    if (isValid.fileExtension && isValid.fileSizeValid) {
-      setClientLogoUrl(URL.createObjectURL(file));
-      setClientLogo(file);
-    } else {
-      if (!isValid.fileExtension && !isValid.fileSizeValid) {
-        setFileUploadError(
-          i18n.t("invalidImageFormatSize", { fileSize: "30" })
-        );
-      } else if (isValid.fileExtension && !isValid.fileSizeValid) {
-        setFileUploadError(i18n.t("invalidImageSize", { fileSize: "30" }));
-      } else {
-        setFileUploadError(i18n.t("invalidImageFormat"));
-      }
-    }
-  };
-
-  const isValidFileUploaded = file => {
-    const validExtensions = ["png", "jpeg", "jpg"];
-    const fileExtensions = file.type.split("/")[1];
-    const validFileByteSize = "30000";
-    const fileSize = file.size;
-
-    return {
-      fileExtension: validExtensions.includes(fileExtensions),
-      fileSizeValid: fileSize <= validFileByteSize,
-    };
-  };
-
   const handleSubmit = async values => {
     const formData = new FormData();
 
@@ -149,7 +84,7 @@ const MobileClientForm = ({
       formData,
       values,
       formType === "new",
-      clientData,
+      client,
       clientLogo,
       clientLogoUrl
     );
@@ -159,116 +94,19 @@ const MobileClientForm = ({
         setClientData([...clientData, { ...res.data, minutes: 0 }]);
         setnewClient(false);
         Toastr.success("Client added successfully");
-      } catch (error) {
-        setApiError(error.message);
+      } catch {
         setSubmitting(false);
       }
     } else {
-      await clientApi
-        .update(clientData.id, formData)
-        .then(() => {
-          setShowEditDialog(false);
-          window.location.reload();
-        })
-        .catch(e => {
-          setApiError(e.message);
-          setSubmitting(false);
-        });
+      try {
+        await clientApi.update(client.id, formData);
+        setShowEditDialog(false);
+        fetchDetails();
+      } catch {
+        setSubmitting(false);
+      }
     }
   };
-
-  const disableBtn = (values, errors) => {
-    if (
-      errors.name ||
-      errors.email ||
-      errors.phone ||
-      errors.address1 ||
-      errors.country ||
-      errors.state ||
-      errors.city ||
-      errors.zipcode ||
-      submitting
-    ) {
-      return true;
-    }
-
-    if (
-      values.name &&
-      values.email &&
-      values.phone &&
-      values.address1 &&
-      values.country &&
-      values.state &&
-      values.city &&
-      values.zipcode
-    ) {
-      return false;
-    }
-
-    return true;
-  };
-
-  const LogoComponent = () => (
-    <div className="my-4 flex flex-row">
-      <div className="mt-2 h-30 w-30 border border-dashed border-miru-dark-purple-400">
-        <div className="profile-img relative m-auto cursor-pointer text-center text-xs font-semibold">
-          <Avatar
-            classNameImg="h-full w-full md:h-full md:w-full"
-            url={clientLogoUrl}
-          />
-          <div className="hover-edit absolute top-0 left-0 h-full w-full bg-miru-white-1000 p-4 opacity-80">
-            <button
-              className="flex flex-row text-miru-han-purple-1000"
-              type="button"
-            >
-              <label className="flex cursor-pointer" htmlFor="file_input">
-                <img
-                  alt="edit"
-                  className="cursor-pointer rounded-full"
-                  src={EditImageButtonSVG}
-                  style={{ minWidth: "40px" }}
-                />
-                <p className="my-auto">Edit</p>
-              </label>
-              <input
-                className="hidden"
-                id="file_input"
-                name="logo"
-                type="file"
-                onChange={onLogoChange}
-              />
-            </button>
-            {clientLogoUrl && (
-              <button
-                className="flex flex-row pl-2 text-miru-red-400"
-                type="button"
-                onClick={handleDeleteLogo}
-              >
-                <img
-                  alt="delete"
-                  src={deleteImageIcon}
-                  style={{ minWidth: "20px" }}
-                />
-                <p className="pl-3">Delete</p>
-              </button>
-            )}
-          </div>
-        </div>
-      </div>
-      <div className="my-auto ml-6 text-xs font-normal text-miru-dark-purple-400">
-        <p>Accepted file formats: PNG and JPG.</p>
-        <p>File size should be &#8826; 30 KB.</p>
-        <p>Image resolution should be 1:1.</p>
-      </div>
-      <input
-        className="hidden"
-        id="file_input"
-        name="logo"
-        type="file"
-        onClick={onLogoChange}
-      />
-    </div>
-  );
 
   return (
     <SidePanel
@@ -277,7 +115,7 @@ const MobileClientForm = ({
     >
       <SidePanel.Header className="mb-2 flex items-center justify-between bg-miru-han-purple-1000 px-5 py-5 text-white lg:bg-white lg:font-bold lg:text-miru-dark-purple-1000">
         <span className="flex w-full items-center justify-center pl-6 text-base font-medium leading-5">
-          {clientData?.id ? "Edit Client" : "Add New Client"}
+          {client?.id ? "Edit Client" : "Add New Client"}
         </span>
         <Button style="ternary" onClick={() => setShowDialog(false)}>
           <XIcon
@@ -288,7 +126,7 @@ const MobileClientForm = ({
       </SidePanel.Header>
       <SidePanel.Body className="sidebar__filters flex h-full flex-col justify-between overflow-y-auto px-4">
         <Formik
-          initialValues={getInitialvalues(clientData)}
+          initialValues={getInitialvalues(client)}
           validationSchema={clientSchema}
           onSubmit={handleSubmit}
         >
@@ -301,38 +139,14 @@ const MobileClientForm = ({
                   <div className="my-4">
                     <div className="field">
                       <div className="mt-1">
-                        {formType == "edit" ? (
-                          <LogoComponent />
-                        ) : clientLogoUrl ? (
-                          <LogoComponent />
-                        ) : (
-                          <div className="mt-2 flex flex-row">
-                            <div className="mt-2 h-30 w-30 border border-dashed border-miru-dark-purple-400 ">
-                              <label
-                                className="flex h-full w-full cursor-pointer justify-center"
-                                htmlFor="file-input"
-                              >
-                                <div className="m-auto cursor-pointer text-center text-xs font-semibold">
-                                  <p className="text-miru-han-purple-1000">
-                                    Select File
-                                  </p>
-                                </div>
-                              </label>
-                              <input
-                                className="hidden"
-                                id="file-input"
-                                name="logo"
-                                type="file"
-                                onChange={onLogoChange}
-                              />
-                            </div>
-                            <div className="my-auto ml-6 text-xs font-normal text-miru-dark-purple-400">
-                              <p>Accepted file formats: PNG and JPG.</p>
-                              <p>File size should be &#8826; 30 KB.</p>
-                              <p>Image resolution should be 1:1.</p>
-                            </div>
-                          </div>
-                        )}
+                        <UploadLogo
+                          clientLogoUrl={clientLogoUrl}
+                          formType={formType}
+                          handleDeleteLogo={handleDeleteLogo}
+                          setClientLogo={setClientLogo}
+                          setClientLogoUrl={setClientLogoUrl}
+                          setFileUploadError={setFileUploadError}
+                        />
                         <p className="mt-3 block max-w-xs text-center text-xs tracking-wider text-red-600">
                           {fileUploadError}
                         </p>
@@ -380,7 +194,7 @@ const MobileClientForm = ({
                           inputClassName="form__input block w-full appearance-none bg-white border-0 focus:border-0 px-0 text-base border-transparent focus:border-transparent focus:ring-0 border-miru-gray-1000 w-full border-bottom-none"
                           name="phone"
                           smartCaret={false}
-                          value={clientData.phone}
+                          value={formType == "edit" ? client.phone : ""}
                           onChange={phone => {
                             setFieldValue("phone", phone);
                           }}
@@ -488,14 +302,11 @@ const MobileClientForm = ({
                     />
                   </div>
                 </div>
-                <p className="mt-3 block text-xs tracking-wider text-red-600">
-                  {apiError}
-                </p>
                 <div className="actions mt-auto">
-                  {clientData?.id ? (
+                  {client?.id ? (
                     <Button
                       className="w-full p-2 text-center text-base font-bold"
-                      disabled={disableBtn(values, errors)}
+                      disabled={disableBtn(values, errors, submitting)}
                       style="primary"
                       type="submit"
                       onClick={handleEdit}
@@ -505,7 +316,7 @@ const MobileClientForm = ({
                   ) : (
                     <Button
                       className="w-full p-2 text-center text-base font-bold"
-                      disabled={disableBtn(values, errors)}
+                      disabled={disableBtn(values, errors, submitting)}
                       style="primary"
                       type="submit"
                       onClick={() => {
@@ -525,5 +336,37 @@ const MobileClientForm = ({
     </SidePanel>
   );
 };
+
+interface IClientForm {
+  client?: any;
+  clientLogoUrl: string;
+  handleDeleteLogo: any;
+  setClientLogoUrl: any;
+  setClientLogo: any;
+  formType?: string;
+  clientData?: any;
+  setClientData?: any;
+  setnewClient?: any;
+  clientLogo?: any;
+  setShowEditDialog?: any;
+  submitting: any;
+  setSubmitting: any;
+  handleEdit?: any;
+  setShowDialog: any;
+  fetchDetails?: any;
+}
+
+interface FormValues {
+  name: string;
+  email: string;
+  phone: string;
+  address1: string;
+  address2: string;
+  country: any;
+  state: any;
+  city: any;
+  zipcode: string;
+  logo: any;
+}
 
 export default MobileClientForm;
