+++ app/javascript/src/components/Clients/Details/constants.ts
@@ -1,0 +1,30 @@
+export const tableHeader = [
+  {
+    Header: "PROJECT",
+    accessor: "col1", // accessor is the "key" in the data
+    cssClass: "md:w-1/3",
+  },
+  {
+    Header: "TEAM",
+    accessor: "col2",
+    cssClass: "md:w-1/3",
+  },
+  {
+    Header: "HOURS LOGGED",
+    accessor: "col3",
+    cssClass: "text-right  md:w-1/5", // accessor is the "key" in the data
+  },
+];
+
+export const mobileTableHeader = [
+  {
+    Header: "PROJECT",
+    accessor: "col1", // accessor is the "key" in the data
+    cssClass: "md:w-1/3",
+  },
+  {
+    Header: "HOURS",
+    accessor: "col2",
+    cssClass: "text-right md:w-1/5", // accessor is the "key" in the data
+  },
+];
